🔨 refactor: 변수삭제 및 값변경
id에 Recipe id 가 들어갈수 있도록 수정

+++ src/main/java/grwm/develop/recipe/dto/SearchRecipe.java
@@ -25,16 +25,20 @@ public class SearchRecipe {
     @AllArgsConstructor
     public static class FindRecipe {
 
-        private int id;
+        private Long id;
         private int reviewCount;
         private float ratingAverage;
         private String title;
         private String image;
         private boolean isPublic;
         private MemberDetail member;
-        public void SetisPublic(boolean isPublic) {this.isPublic = isPublic;}
+
+        public void SetisPublic(boolean isPublic) {
+            this.isPublic = isPublic;
+        }
 
     }
+
     @Getter
     @AllArgsConstructor
     public static class MemberDetail {
